The basics to implement #4 are in place :sunglasses:
